+++ Source/DafnyLanguageServer/Language/IProgramVerifier.cs
@@ -26,7 +26,7 @@ namespace Microsoft.Dafny.LanguageServer.Language {
     /// <returns>The result of the verification run.</returns>
     /// <exception cref="System.OperationCanceledException">Thrown when the cancellation was requested before completion.</exception>
     /// <exception cref="System.ObjectDisposedException">Thrown if the cancellation token was disposed before the completion.</exception>
-    Task<IReadOnlyList<IImplementationTask>> GetVerificationTasksAsync(CompilationAfterResolution document,
+    Task<IReadOnlyList<IImplementationTask>> GetVerificationTasksAsync(DocumentAfterResolution document,
       IReadOnlyDictionary<Position, int> implementationOrder, CancellationToken cancellationToken);
     public IObservable<AssertionBatchResult> BatchCompletions { get; }
   }
