+++ Source/DafnyLanguageServer/Workspace/DocumentManagerDatabase.cs
@@ -72,25 +72,25 @@ namespace Microsoft.Dafny.LanguageServer.Workspace {
       return false;
     }
 
-    public Task<CompilationView?> GetResolvedDocumentAsync(TextDocumentIdentifier documentId) {
+    public Task<DocumentSnapshot?> GetResolvedDocumentAsync(TextDocumentIdentifier documentId) {
       if (documents.TryGetValue(documentId.Uri, out var state)) {
-        return state.GetResolvedDocumentAsync();
+        return state.GetSnapshotAfterResolutionAsync();
       }
-      return Task.FromResult<CompilationView?>(null);
+      return Task.FromResult<DocumentSnapshot?>(null);
     }
 
-    public Task<CompilationAfterParsing?> GetLastDocumentAsync(TextDocumentIdentifier documentId) {
+    public Task<DocumentAfterParsing?> GetLastDocumentAsync(TextDocumentIdentifier documentId) {
       if (documents.TryGetValue(documentId.Uri, out var databaseEntry)) {
         return databaseEntry.LastDocumentAsync!;
       }
-      return Task.FromResult<CompilationAfterParsing?>(null);
+      return Task.FromResult<DocumentAfterParsing?>(null);
     }
 
     public DocumentManager? GetDocumentManager(TextDocumentIdentifier documentId) {
       return documents.GetValueOrDefault(documentId.Uri);
     }
 
-    public IEnumerable<CompilationManager> Documents => documents.Values.Select(m => m.CompilationManager);
+    public IEnumerable<Compilation> Documents => documents.Values.Select(m => m.Compilation);
 
   }
 }
