+++ Source/DafnyLanguageServer/Workspace/IDocumentDatabase.cs
@@ -47,17 +47,17 @@ namespace Microsoft.Dafny.LanguageServer.Workspace {
     /// </summary>
     /// <param name="documentId">The ID of the document to resolve.</param>
     /// <returns>An instance of the managed document, <c>null</c> if the specified document was not found.</returns>
-    Task<CompilationView?> GetResolvedDocumentAsync(TextDocumentIdentifier documentId);
+    Task<DocumentSnapshot?> GetResolvedDocumentAsync(TextDocumentIdentifier documentId);
 
     /// <summary>
     /// Tries to resolve a verified document with the specified identifier.
     /// </summary>
     /// <param name="documentId">The ID of the document to resolve.</param>
     /// <returns>An instance of the managed document, <c>null</c> if the specified document was not found.</returns>
-    Task<CompilationAfterParsing?> GetLastDocumentAsync(TextDocumentIdentifier documentId);
+    Task<DocumentAfterParsing?> GetLastDocumentAsync(TextDocumentIdentifier documentId);
 
     DocumentManager? GetDocumentManager(TextDocumentIdentifier documentId);
 
-    IEnumerable<CompilationManager> Documents { get; }
+    IEnumerable<Compilation> Documents { get; }
   }
 }
