+++ Source/DafnyLanguageServer/Workspace/IDocumentEntry.cs
@@ -3,10 +3,10 @@ using System.Threading.Tasks;
 namespace Microsoft.Dafny.LanguageServer.Workspace;
 
 public interface IDocumentEntry {
-  Task<Compilation> ResolvedDocument { get; }
-  Task<Compilation> TranslatedDocument { get; }
-  Compilation LastPublishedDocument { get; }
-  Task<Compilation> LastDocument { get; }
+  Task<Document> ResolvedDocument { get; }
+  Task<Document> TranslatedDocument { get; }
+  Document LastPublishedDocument { get; }
+  Task<Document> LastDocument { get; }
   public bool Idle { get; }
   void MarkVerificationStarted();
