+++ Source/DafnyLanguageServer/Workspace/INotificationPublisher.cs
@@ -4,19 +4,19 @@ using OmniSharp.Extensions.LanguageServer.Protocol.Models;
 namespace Microsoft.Dafny.LanguageServer.Workspace {
   /// <summary>
   /// Implementations of this interface are responsible to publish the diagnostics
-  /// of a <see cref="Compilation"/> to the LSP client.
+  /// of a <see cref="Document"/> to the LSP client.
   /// </summary>
   public interface INotificationPublisher {
     /// <summary>
     /// Publishes the diagnostics of the specified dafny document to the connected LSP client.
     /// </summary>
     /// <param name="document">The document whose diagnostics should be published.</param>
-    void PublishNotifications(CompilationView previous, CompilationView document);
+    void PublishNotifications(DocumentSnapshot previous, DocumentSnapshot document);
 
     /// <summary>
     /// Publishes the more precise real-time verification diagnostics to the connected LSP client
     /// </summary>
-    void PublishGutterIcons(CompilationView document, bool verificationStarted);
+    void PublishGutterIcons(DocumentSnapshot document, bool verificationStarted);
 
     /// <summary>
     /// Hides the previously published diagnostics of the specified dafny document.
