+++ Source/DafnyLanguageServer/Workspace/ITelemetryPublisher.cs
@@ -5,7 +5,7 @@ using OmniSharp.Extensions.LanguageServer.Protocol.Models;
 namespace Microsoft.Dafny.LanguageServer.Workspace {
   /// <summary>
   /// Implementations of this interface are responsible to publish telemetry events
-  /// of a <see cref="Compilation"/> to the LSP client.
+  /// of a <see cref="Document"/> to the LSP client.
   /// </summary>
   public interface ITelemetryPublisher {
     protected enum TelemetryEventKind {
