+++ Source/DafnyLanguageServer/Workspace/ITextDocumentLoader.cs
@@ -19,7 +19,7 @@ namespace Microsoft.Dafny.LanguageServer.Workspace {
     /// <returns>The unloaded dafny document.</returns>
     /// <exception cref="System.OperationCanceledException">Thrown when the cancellation was requested before completion.</exception>
     /// <exception cref="System.ObjectDisposedException">Thrown if the cancellation token was disposed before the completion.</exception>
-    CompilationView CreateUnloaded(DocumentTextBuffer textDocument, CancellationToken cancellationToken);
+    DocumentSnapshot CreateUnloaded(DocumentTextBuffer textDocument, CancellationToken cancellationToken);
 
 
     /// <summary>
@@ -31,6 +31,6 @@ namespace Microsoft.Dafny.LanguageServer.Workspace {
     /// <returns>The loaded dafny document.</returns>
     /// <exception cref="System.OperationCanceledException">Thrown when the cancellation was requested before completion.</exception>
     /// <exception cref="System.ObjectDisposedException">Thrown if the cancellation token was disposed before the completion.</exception>
-    Task<CompilationAfterParsing> LoadAsync(DocumentTextBuffer textDocument, CancellationToken cancellationToken);
+    Task<DocumentAfterParsing> LoadAsync(DocumentTextBuffer textDocument, CancellationToken cancellationToken);
   }
 }
